+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSTemplesConfig.java
@@ -115,10 +115,10 @@ public class RSTemplesConfig implements ConfigData {
 
         @ConfigEntry.Gui.Tooltip(count = 0)
         @ConfigEntry.Gui.PrefixText
-        @Comment("How rare are Badlands Pyramid in non-plateau Badlands biomes."
+        @Comment("How rare are Badlands Pyramid in Badlands biomes."
                 + "\n1 for spawning in most chunks and 1001 for none.")
         @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-        public int badlandsPyramidMaxChunkDistance = 20;
+        public int badlandsPyramidMaxChunkDistance = 25;
 
         @ConfigEntry.Gui.Tooltip(count = 0)
         @ConfigEntry.Gui.PrefixText