+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSVillagesConfig.java
@@ -27,49 +27,49 @@ public class RSVillagesConfig implements ConfigData {
     @Comment("How rare are Badlands Villages in Badland biomes."
             + "\n1 for spawning in most chunks and 1001 for no spawn.")
     @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-    public int badlandsVillageMaxChunkDistance = 23;
+    public int badlandsVillageMaxChunkDistance = 30;
     
     @ConfigEntry.Gui.Tooltip(count = 0)
     @ConfigEntry.Gui.PrefixText
     @Comment("How rare are Birch Villages in Birch biomes."
             + "\n1 for spawning in most chunks and 1001 for no spawn.")
     @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-    public int birchVillageMaxChunkDistance = 31;
+    public int birchVillageMaxChunkDistance = 38;
 
     @ConfigEntry.Gui.Tooltip(count = 0)
     @ConfigEntry.Gui.PrefixText
     @Comment("How rare are Dark Forest Villages in Dark Forest biomes."
             + "\n1 for spawning in most chunks and 1001 for no spawn.")
     @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-    public int darkForestVillageMaxChunkDistance = 31;
+    public int darkForestVillageMaxChunkDistance = 38;
 
     @ConfigEntry.Gui.Tooltip(count = 0)
     @ConfigEntry.Gui.PrefixText
     @Comment("How rare are Jungle Villages in Jungle biomes."
             + "\n1 for spawning in most chunks and 1001 for no spawn.")
     @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-    public int jungleVillageMaxChunkDistance = 29;
+    public int jungleVillageMaxChunkDistance = 38;
 
     @ConfigEntry.Gui.Tooltip(count = 0)
     @ConfigEntry.Gui.PrefixText
     @Comment("How rare are Swamp Villages in Swamp biomes."
             + "\n1 for spawning in most chunks and 1001 for no spawn.")
     @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-    public int swampVillageMaxChunkDistance = 31;
+    public int swampVillageMaxChunkDistance = 38;
 
     @ConfigEntry.Gui.Tooltip(count = 0)
     @ConfigEntry.Gui.PrefixText
     @Comment("How rare are Mountains Villages in Mountains biomes."
             + "\n1 for spawning in most chunks and 1001 for no spawn.")
     @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-    public int mountainsVillageMaxChunkDistance = 31;
+    public int mountainsVillageMaxChunkDistance = 38;
 
     @ConfigEntry.Gui.Tooltip(count = 0)
     @ConfigEntry.Gui.PrefixText
     @Comment("How rare are Giant Taiga Villages in Giant Taiga biomes."
             + "\n1 for spawning in most chunks and 1001 for no spawn.")
     @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-    public int giantTaigaVillageMaxChunkDistance = 29;
+    public int giantTaigaVillageMaxChunkDistance = 38;
 
 
     @ConfigEntry.Gui.Tooltip(count = 0)
@@ -91,5 +91,5 @@ public class RSVillagesConfig implements ConfigData {
     @Comment("How rare are Oak Villages in forest category"
             + "\nbiomes that are not birch or dark forest.")
     @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-    public int villageOakMaxChunkDistance = 33;
+    public int villageOakMaxChunkDistance = 38;
 }